Making changes

+++ Exam-Prep/Exam_16-Aug-20/skeleton/project/hardware/power_hardware.py
@@ -5,4 +5,4 @@ class PowerHardware(Hardware):
     TYPE = "Power"
 
     def __init__(self, name, capacity, memory):
-        super().__init__(name, self.TYPE, capacity * 0.25, memory * 1.75)
+        super().__init__(name, self.TYPE, int(capacity * 0.25), int(memory * 1.75))
